// src/app/main/reports/report-type.ts
export type DurationType = 'month' | 'week' | 'day';
// ...
export type FilterState = {
  startDate: moment.Moment;
  endDate: moment.Moment;
  duration?: DurationType;
  group1?: string;
  group2?: string;
  machines?: number[];
  downtimeCodes?: number[];
  delayCodes?: number[];
  shifts?: number[];
  scheduleStatus?: string;
  isBack?: boolean;
};

export type CoilScrapReport = {
  coil: Coil;
  material: IMaterialDoc & { documentID: string };
  groups: {
    groupTitle: 'Exempt' | 'Non-Exempt';
    reasons: {
      scrapCode: number;
      scrapReason: string;
      scrapFt: number;
      scrapPct: number;
    }[];
    scrapFt: number;
    scrapPct: number;
  };
};

export type OrderSummary = {
  orderCode: string;
  materialCode: string;
  toolingCode: string;
  goodFt: number;
  goodLbs: number;
  scrapFt: number;
  scrapLbs: number;
  reclaimedScrapFt: number;
  bundles: {
    bundleIdentity: string;
    goodFt: number;
    goodLbs: number;
    scrapFt: number;
    scrapLbs: number;
    reclaimedScrapFt: number;
    prodRuns: {
      coilSerialNumber: string;
      machineNumber: number;
      productionDate: Date;
      shift: number;
      employeeNumber: string;
      goodFt: number;
      goodLbs: number;
      scrapFt: number;
      scrapLbs: number;
      reclaimedScrapFt: number;
      items: {
        bundle: number;
        partLengthIn: number;
        punchPartName: string;
        punchOption: string;
        externalItemId: string;
        quantity: number;
        goodFt: number;
        scrapFt: number;
        reclaimedScrapFt: number;
      }[];
    }[];
  }[];
};